Apply idiomatic Rust cleanups
- Replace many to_string()/into() clones with to_owned(), take, or
  direct moves
- Use is_multiple_of, is_some_and, map(String::as_str) where appropriate
- Simplify control flow (if let, reversed branches, early returns) and
  minor formatting fixes

+++ orion-lib/src/listeners/tcp_proxy.rs
@@ -85,7 +85,7 @@ impl TcpProxyBuilder {
     #[inline]
     pub fn build(self) -> Result<TcpProxy> {
         let listener_name = self.listener_name.unwrap_or("listener name is not set");
-        let filterchain_id = self.filterchain_id.unwrap_or(0 as u64);
+        let filterchain_id = self.filterchain_id.unwrap_or(0_u64);
         let TcpProxyConfig { cluster_specifier, access_log } = self.tcp_proxy_config;
         Ok(TcpProxy { listener_name, filterchain_id, access_log, cluster: cluster_specifier })
     }
